requirements.txt
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.16.0
dash-bootstrap-components==0.10.5
dash-core-components==1.12.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.0
dash-table==4.10.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
gspread==3.6.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multielo @ git+git://github.com/djcunningham0/multielo.git@9113470b33753fae59a229bd17f80296df0ae130
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.2
plotly==4.9.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
